update refactor the code
